move PlotArea, DataManager and TableManager to spec folder
